🚚 user 폴더 app 폴더 내부로 경로 수정
