+++ src/app/user/entity/batting-type.ts
@@ -1,0 +1,6 @@
+export const BattingType = {
+  LEFT: 1,
+  RIGHT: 2,
+} as const;
+
+export type BattingType = (typeof BattingType)[keyof typeof BattingType];
