+++ src/app/user/entity/pitching-type.ts
@@ -1,0 +1,6 @@
+export const PitchingType = {
+  LEFT: 1,
+  RIGHT: 2,
+} as const;
+
+export type PitchingType = (typeof PitchingType)[keyof typeof PitchingType];
