+++ src/app/user/entity/sex.ts
@@ -1,0 +1,6 @@
+export const Sex = {
+  MALE: 1,
+  FEMALE: 2,
+} as const;
+
+export type Sex = (typeof Sex)[keyof typeof Sex];
