--- src/user/entity/sex.ts
@@ -1,6 +1,0 @@
-export const Sex = {
-  MALE: 1,
-  FEMALE: 2,
-} as const;
-
-export type Sex = (typeof Sex)[keyof typeof Sex];
