--- src/user/user.module.ts
@@ -1,8 +1,0 @@
-import { Module } from '@nestjs/common';
-
-@Module({
-  imports: [],
-  controllers: [],
-  providers: [],
-})
-export class UserModule {}
